+++ src/TestTrackingDiagrams.Extensions.Elasticsearch/ElasticsearchTrackingVerbosity.cs
@@ -1,0 +1,3 @@
+namespace TestTrackingDiagrams.Extensions.Elasticsearch;
+
+public enum ElasticsearchTrackingVerbosity { Raw, Detailed, Summarised }